Add Standby Screen, Startup Screen and WiFi Utils

+++ src/modules/wifi/evilportal.cpp
@@ -140,7 +140,7 @@ bool EvilPortalAddtional::apSetup(bool _deauth) {
         if (access_points->get(i).selected) {
             if (_deauth) {
                 memcpy(deauth_frame, wifi_obj.deauth_frame_packet, sizeof(wifi_obj.deauth_frame_packet));
-                esp_wifi_set_channel((int)access_points->get(i).channel, WIFI_SECOND_CHAN_NONE);
+                esp_wifi_set_channel(access_points->get(i).channel, WIFI_SECOND_CHAN_NONE);
                 vTaskDelay(50 / portTICK_PERIOD_MS);
                 deauth_frame[10] = access_points->get(i).bssid[0];
                 deauth_frame[11] = access_points->get(i).bssid[1];
